DB fix

+++ Lab/Lab/Models/Demand.cs
@@ -3,7 +3,7 @@
 		public int Id { get; set; }
 		public Test Test { get; set; }
 		public string DemandName { get; set; }
-		public int LowerBorder { get; set; }
-		public int UpperBorder { get; set; }
+		public float LowerBorder { get; set; }
+		public float UpperBorder { get; set; }
 	}
 }
